backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic[email]>=2.0.0
pymongo>=4.0.0
python-dotenv>=1.0.0
pydantic-settings>=1.0.0
python-jose[cryptography]>=3.3.0
python-socketio[asyncio_server]>=5.10.0
python-socketio[client]>=5.10.0
websockets>=10.0
httpx>=0.28.0
requests>=2.20.0
motor>=3.0.0
fastapi-mail>=1.4.1
mangum>=0.17.0